unify history handling
