+++ features/omni-kit/controllers/borrow/OmniBorrowFormController.tsx
@@ -42,14 +42,7 @@ export function OmniBorrowFormController({ txHandler }: { txHandler: () => () =>
       elements: { riskSidebar },
     },
   } = useOmniProductContext(OmniProductType.Borrow)
-  console.log(
-    'supportedMultiplyTokens',
-    isPoolSupportingMultiply({
-      collateralToken,
-      quoteToken,
-      supportedTokens: supportedMultiplyTokens[networkId],
-    }),
-  )
+
   return (
     <OmniFormView
       {...(!isOpening && {
