+++ features/omni-kit/protocols/ajna/history/ajnaDefaultHistoryEvent.ts
@@ -1,0 +1,17 @@
+import { unifiedDefaultHistoryItem } from 'features/positionHistory/consts'
+import { zero } from 'helpers/zero'
+
+export const ajnaDefaultHistoryEvent = {
+  ...unifiedDefaultHistoryItem,
+  quoteTokensMoved: zero,
+  moveQuoteFromPrice: zero,
+  moveQuoteToPrice: zero,
+  addOrRemovePrice: zero,
+  originationFee: zero,
+  originationFeeInQuoteToken: zero,
+  settledDebt: zero,
+  debtToCover: zero,
+  collateralForLiquidation: zero,
+  remainingCollateral: zero,
+  auction: undefined,
+}
