+++ features/omni-kit/protocols/morpho-blue/history/index.ts
@@ -1,1 +1,1 @@
-export * from './mapLendingEvents'
+export * from './mapMorphoLendingEvents'
