+++ features/omni-kit/protocols/morpho-blue/history/types.ts
@@ -1,0 +1,3 @@
+import type { PositionHistoryEvent } from 'features/positionHistory/types'
+
+export type MorphoHistoryEvent = PositionHistoryEvent
